--- angular/src/app/dialogs/column-bibliography.html
@@ -1,5 +1,0 @@
-<!-- This dialog simply shows whatever input has been provided -->
-<div mat-dialog-content>
-  <h3 style="padding-bottom: 0.2em">Column Bibliography</h3>
-  <div [innerHTML]="data.content | safeHtml"></div>
-</div>
