--- angular/src/app/dialogs/custom-dialog.html
@@ -1,4 +1,0 @@
-<!-- This dialog simply shows whatever input has been provided -->
-<div mat-dialog-content>
-  <div [innerHTML]="data.content | safeHtml"></div>
-</div>
